[regression] Make the two audited-fix tests pass on Windows
github_1932 cast argv[0] through unsigned long, which is 32-bit under
LLP64, so the guard admitted non-null pointers and the assertion was
genuinely refutable there; uintptr_t is lossless on every data model.
github_2163 reports the assertion UNREACHED on Windows, where MSVC's
__security_cookie has no definition, so mark it THOROUGH until that gap
is modelled.

Refs #6954

+++ regression/esbmc/github_1932/main.c
@@ -1,9 +1,12 @@
+// The round trip has to be through a pointer-sized integer: `unsigned long`
+// is 32-bit under LLP64, so the cast truncated argv[0] and the guard admitted
+// non-null pointers whose low word happened to be zero. #1932
 #include <assert.h>
+#include <stdint.h>
 
 int main(int argc, char **argv) {
-  if ((unsigned long)argv[0] == (unsigned long)0) {
+  if ((uintptr_t)argv[0] == (uintptr_t)0) {
     assert(argv[0] == 0);
   }
   return 0;
 }
-
